+++ docs/src/dev-docs/changelog.rst
@@ -7,9 +7,6 @@ All notable changes to ``metatrain`` are documented here, following the `keep a
 changelog <https://keepachangelog.com/en/1.1.0/>`_ format. This project follows
 `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-Unreleased
-----------
-
 .. Possible sections for each release:
 
 .. Fixed
@@ -23,6 +20,16 @@ Unreleased
 
 .. Removed
 .. #######
+
+Unreleased
+----------
+
+Changed
+#######
+
+- :func:`metatrain.util.io.load_model` does not copy a remote model to the current
+  directory.
+
 
 Version 2025.2 - 2025-03-11
 ---------------------------